FEPOLY LaTeX Files
The first readable output for the FEPOLY project by Amir Parvardi. Mohsen Katebi helped me with the Python code for extracting the problems from the AoPS math olympiad forums. There are a ton of problems on functional equations and polynomials in this project. Feel free to use however you want.